Update pattern matching exhaustivity check files
The diffs in t6420 and t7466 comes from a bug a the previous algorithm
that would conclue that given `val a: Int`, `1` and `a.type` are non-
intersecting because they are both singleton types. However, we can only
get to that conclusion when both types are ConstantTypes.

The diff in andtype-opentype-interaction come from the fact that the new
decomposition of sealed types is more powerful than the previous one.
For instance, we can now rule out `SealedClass & OpenTrait` as both
types are provably non intersecting.